"Refactor ESLint configuration and update dependencies"
The `.eslintrc` file has been deleted and replaced with `.eslintrc.cjs`. This enhances ESLint configuration by setting specific environments and extends standard-with-typescript which unifies JavaScript and TypeScript linting rules.

`package.json` and `package-lock.json` are updated to reflect removed or updated dependencies. Among the updates, `@tauri-apps/api` was added as a new devDependency, and some unused packages related to TypeScript ESLint were removed, streamlining the list of dependencies.

Note: All this was done in a bid to clean up the codebase and ensure that all linting rules are appropriately configured while removing unnecessary dependencies.