Add anti-piracy check
Added module noPiracy.ts to redirect non-valid domains in non-development environments. This anti-piracy check ensures that the application is run only on approved domains and redirects others to the valid site's piracy page. The function is subsequently imported and called in the main index.js file. This change is crucial in deterring unauthorized distribution and usage of the application.