Move build scripts to 'build' directory and refactor
This commit includes moving all build-related scripts from 'scripts' to a new 'build' directory to better separate concerns and improve code organization. The refactoring involves simplifying the main build.js script by extracting its build options into a new standalone esbuild.js script in a new 'meta' sub-directory. This arrangement increases the readability and maintainability of the build scripts. The esbuild() calls in tasks/kernel.js are also updated to use the extracted build options. Script paths in package.json are updated accordingly for consistency. The renaming and relocating of other scripts are to align with this update and to make the scripts structure more logical and systematic.