Add CSS loader and minify HTML for improving performance
This commit adds CSS as a loader type in the esbuild.js file. This allows the application to treat CSS files as modules and include them in the build. A new dependency, minify-html package, was also added to the application. The purpose of this package is to minify HTML files at build-time, reducing file sizes and thereby improving the speed and performance.