Refactor project from JavaScript to TypeScript
This commit represents a significant overhaul of our codebase from JavaScript to TypeScript to leverage static typing and improve code quality. The diff shows changes across the entire codebase, including restructured directory and renaming of extension from .js to .ts in relevant areas. This change also enhances error handling mechanism in loadInstaller function and introduces a write method for better logging. Additionally, there is a bumped version of the application with changes from open-source to closed-source in the package.json file.